fix(c-workshop: fixed examples, makefile, and gitignore

+++ Examples/example1.c
@@ -4,14 +4,14 @@
 
 int main(void) {
 
-   printf("size of a char: %d bytes\n", sizeof(char) );
-   printf("size of a int: %d bytes\n", sizeof(int));
-   printf("size of a short: %d bytes\n", sizeof(short));
-   printf("size of a long: %d bytes\n", sizeof(long));
-   printf("size of a float: %d bytes\n", sizeof(float));
-   printf("size of a double: %d bytes\n", sizeof(double));
-   printf("size of a int16_t: %d bytes\n", sizeof(int16_t));
-   printf("size of a int32_t: %d bytes\n", sizeof(int32_t));
+   printf("size of a char: %d bytes\n", (int) sizeof(char) );
+   printf("size of a int: %d bytes\n", (int) sizeof(int));
+   printf("size of a short: %d bytes\n", (int) sizeof(short));
+   printf("size of a long: %d bytes\n", (int) sizeof(long));
+   printf("size of a float: %d bytes\n", (int) sizeof(float));
+   printf("size of a double: %d bytes\n", (int) sizeof(double));
+   printf("size of a int16_t: %d bytes\n", (int) sizeof(int16_t));
+   printf("size of a int32_t: %d bytes\n", (int) sizeof(int32_t));
 
    return(EXIT_SUCCESS);
 }
